moved admin_required into journalist_app/decorators.py
